chore: Move all code to index.ts
