--- src/SASjsConfig.ts
@@ -1,8 +1,0 @@
-export class SASjsConfig {
-  baseURL: string = "";
-  port: number | null = null;
-  pathSAS9: string = "";
-  pathSASViya: string = "";
-  programRoot: string = "";
-  serverType: string = "";
-}
