--- src/SASjsRequest.ts
@@ -1,9 +1,0 @@
-import * as moment from "moment";
-
-export interface SASjsRequest {
-  serviceLink: string;
-  timestamp: moment.Moment;
-  sourceCode: string;
-  generatedCode: string;
-  logLink: string;
-}
